requirements.txt
flask==1.1.4
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1
pillow==9.0.0
pillow-heif==0.10.0
requests==2.31.0
PyYAML==6.0.1
piexif==1.1.3